Moving util.rb out of Object. #1
Removes uneccessary init functions.
Adds beginning of test suite.
