+++ Rakefile
@@ -1,2 +1,10 @@
 require "bundler/gem_tasks"
+require 'rake/testtask'
 
+Rake::TestTask.new do |t|
+  t.libs << 'test'
+  t.pattern = "test/**/test*.rb"
+end
+
+desc "Run tests"
+task :default => :test
