+++ lib/marathon.rb
@@ -4,6 +4,7 @@ require 'net/http'
 require 'uri'
 
 require 'config'
+require "orch/util"
 
 class String
   def numeric?
@@ -13,8 +14,8 @@ end
 
 module Orch
   class Marathon
-    def initialize(options)
-    end
+
+    include Orch::Util
 
     def deploy(url_list, app_id, json_payload)
       if url_list.nil?
